update: querys
