+++ packages/graphql-client/documents/token-holder/GetDaoAddressesAccountBalances.graphql
